Clase 186: Mostrar mensaje de confirmacion y espera

+++ src/modules/daybook/views/EntryView.vue
@@ -37,6 +37,7 @@
 <script>
 import { defineAsyncComponent } from "vue"
 import { mapGetters, mapActions } from "vuex" // computed
+import Swal from "sweetalert2"
 
 import getDayMonthYear from "../helpers/getDayMonthYear"
 
@@ -88,22 +89,42 @@ export default {
       this.entry = entry
     },
     async saveEntry() {
+      new Swal({
+        title: "Espere por favor",
+        allowOutsideClick: false,
+      })
+      Swal.showLoading()
+
       if (this.entry.id) {
         // Actualizar
         await this.updateEntry(this.entry)
       } else {
         // Crear una nueva entrada
         const id = await this.createEntry(this.entry)
-
         this.$router.push({ name: "entry", params: { id } })
       }
+
+      Swal.fire("Guardado", "Entrada registrada con éxito", "success")
     },
     async onDeleteEntry() {
-      console.log("delete", this.entry)
+      const { isConfirmed } = await Swal.fire({
+        title: "Está seguro?",
+        text: "Una vez borrado, no se puede recuperar",
+        showDenyButton: true,
+        confirmButtonText: "Sí estoy seguro",
+      })
 
-      await this.deleteEntry(this.entry.id)
+      if (isConfirmed) {
+        new Swal({
+          title: "Espere por favor",
+          allowOutsideClick: false,
+        })
+        Swal.showLoading()
+        await this.deleteEntry(this.entry.id)
+        this.$router.push({ name: "no-entry" })
 
-      this.$router.push({ name: "no-entry" })
+        Swal.fire("Eliminado", "", "success")
+      }
     },
   },
   created() {
